requirements.txt
TinyDB
ujson
discord.py
